requirements.txt
metaapi-cloud-sdk>=28.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1